@@ -2,81 +2,50 @@
   "name": "starships",
   "version": "0.1.0",
   "private": true,
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "documentation": "node scripts/documentation.mjs"
+  },
   "dependencies": {
-    "@auth0/auth0-react": "^1.9.0",
+    "@auth0/auth0-react": "^1.10.1",
     "@monaco-editor/react": "^4.2.2",
-    "@sentry/react": "^6.19.4",
-    "@sentry/tracing": "^6.19.4",
+    "@sentry/react": "^6.19.7",
+    "@sentry/tracing": "^6.19.7",
     "@svb-41/core": "^0.14.0",
     "@svb-41/engine": "^0.12.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^27.0.1",
-    "@types/node": "^16.4.13",
-    "@types/react": "^17.0.17",
-    "@types/react-dom": "^17.0.9",
-    "craco": "^0.0.3",
-    "craco-alias": "^3.0.1",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^14.2.0",
+    "@types/jest": "^27.5.1",
+    "@types/node": "^17.0.35",
+    "@types/prettier": "^2.6.1",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.5",
     "jwt-decode": "^3.1.2",
     "monaco-themes": "^0.4.0",
-    "pixi.js": "^6.1.2",
-    "prettier": "^2.3.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "pixi.js": "^6.3.2",
+    "prettier": "^2.6.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-markdown": "^8.0.1",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "4.0.3",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
     "react-spring": "^9.4.5",
-    "redux": "^4.1.2",
+    "redux": "^4.2.0",
     "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^2.4.1",
     "rehype-highlight": "^5.0.2",
     "sass": "^1.49.9",
-    "typescript": "^4.3.5",
+    "typescript": "^4.7.2",
     "uuid": "^8.3.2",
-    "web-vitals": "^1.0.1"
-  },
-  "scripts": {
-    "start": "craco start",
-    "build": "CI=;craco build",
-    "test": "craco test",
-    "eject": "craco eject",
-    "documentation": "node scripts/documentation.mjs"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ],
-    "rules": {
-      "semi": 0,
-      "no-shadow": 0,
-      "curly": 0,
-      "no-unused-vars": 0,
-      "@typescript-eslint/no-unused-vars": 0,
-      "@typescript-eslint/no-redeclare": 0,
-      "react-hooks/exhaustive-deps": 0
-    }
-  },
-  "resolutions": {
-    "react-error-overlay": "6.0.9"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "web-vitals": "^2.1.4"
   },
   "devDependencies": {
-    "@types/uuid": "^8.3.1",
-    "node-fetch": "^3.2.2"
+    "@types/uuid": "^8.3.4",
+    "@vitejs/plugin-react": "^1.3.2",
+    "node-fetch": "^3.2.4",
+    "vite": "^2.9.9"
   }
 }
